feat(passport): add client scan wrapper and shared outcome types
