+++ packages/common/src/types/api/index.ts
@@ -1,3 +1,4 @@
 export * from "./application";
 export * from "./file";
 export * from "./github";
+export * from "./passport";
